+++ RELEASES.md
@@ -63,3 +63,83 @@ Based on:
 - Speakeasy CLI 1.19.2 (2.16.5) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [NPM v0.8.2] https://www.npmjs.com/package/@speakeasy-sdks/sensible/v/0.8.2 - .
+
+## 2023-04-01 00:46:28
+### Changes
+Based on:
+- OpenAPI Doc 0.0.0 ./sensible.yml
+- Speakeasy CLI 1.19.3 (2.16.7) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v0.8.3] https://www.npmjs.com/package/@speakeasy-sdks/sensible/v/0.8.3 - .
+
+## 2023-04-06 00:42:53
+### Changes
+Based on:
+- OpenAPI Doc 0.0.0 ./sensible.yml
+- Speakeasy CLI 1.19.6 (2.17.8) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v0.9.0] https://www.npmjs.com/package/@speakeasy-sdks/sensible/v/0.9.0 - .
+
+## 2023-04-12 00:43:06
+### Changes
+Based on:
+- OpenAPI Doc 0.0.0 ./sensible.yml
+- Speakeasy CLI 1.19.7 (2.17.9) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v0.9.1] https://www.npmjs.com/package/@speakeasy-sdks/sensible/v/0.9.1 - .
+
+## 2023-04-14 00:43:05
+### Changes
+Based on:
+- OpenAPI Doc 0.0.0 ./sensible.yml
+- Speakeasy CLI 1.20.0 (2.18.0) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v0.10.0] https://www.npmjs.com/package/@speakeasy-sdks/sensible/v/0.10.0 - .
+
+## 2023-04-18 00:44:11
+### Changes
+Based on:
+- OpenAPI Doc 0.0.0 ./sensible.yml
+- Speakeasy CLI 1.20.1 (2.18.1) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v0.10.1] https://www.npmjs.com/package/@speakeasy-sdks/sensible/v/0.10.1 - .
+
+## 2023-04-19 00:45:00
+### Changes
+Based on:
+- OpenAPI Doc 0.0.0 ./sensible.yml
+- Speakeasy CLI 1.20.2 (2.18.2) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v0.10.2] https://www.npmjs.com/package/@speakeasy-sdks/sensible/v/0.10.2 - .
+
+## 2023-04-21 00:43:15
+### Changes
+Based on:
+- OpenAPI Doc 0.0.0 ./sensible.yml
+- Speakeasy CLI 1.21.4 (2.19.1) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v0.11.0] https://www.npmjs.com/package/@speakeasy-sdks/sensible/v/0.11.0 - .
+
+## 2023-04-22 00:44:43
+### Changes
+Based on:
+- OpenAPI Doc 0.0.0 ./sensible.yml
+- Speakeasy CLI 1.22.1 (2.20.1) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v0.12.0] https://www.npmjs.com/package/@speakeasy-sdks/sensible/v/0.12.0 - .
+
+## 2023-04-26 00:44:19
+### Changes
+Based on:
+- OpenAPI Doc 0.0.0 ./sensible.yml
+- Speakeasy CLI 1.23.1 (2.21.1) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v0.13.0] https://www.npmjs.com/package/@speakeasy-sdks/sensible/v/0.13.0 - .
+
+## 2023-04-26 08:30:03
+### Changes
+Based on:
+- OpenAPI Doc 0.0.0 
+- Speakeasy CLI 1.23.1 (2.21.1) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v0.13.1] https://www.npmjs.com/package/@speakeasy-sdks/sensible/v/0.13.1 - .